# docs/swagger.yaml
definitions:
...
  server.passkeyLoginOptionsRequest:
    properties:
      email:
        type: string
    type: object
...
  server.updatePasswordRequest:
    properties:
      current_password:
        type: string
      password:
        type: string
      password_confirmation:
        type: string
    type: object
  server.updateProfileRequest:
    properties:
      email:
        type: string
      name:
        type: string
    type: object
...
host: localhost:4000
info:
  contact: {}
  description: REST API aplikasi EBITDA Max APN (refactor Go + Next.js) — scope manager
    KDKMP & superadmin.
  title: EBITDA Max APN API
  version: "1.0"
paths:
...
  /api/v1/auth/passkey/login:
    post:
      consumes:
      - application/json
      description: Memverifikasi respons assertion dan membuat session login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Selesaikan login passkey
      tags:
      - Auth
  /api/v1/auth/passkey/options:
    post:
      consumes:
      - application/json
      description: Mengembalikan opsi WebAuthn (publicKey) untuk login passkey berdasarkan
        email; menyimpan challenge di cookie `ebitda_passkey`.
      parameters:
      - description: Email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.passkeyLoginOptionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mulai login passkey
      tags:
      - Auth
...
  /api/v1/passkeys:
    get:
      description: Menampilkan passkey milik user yang sedang login.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Daftar passkey
      tags:
      - Passkeys
  /api/v1/passkeys/{id}:
    delete:
      description: Menghapus passkey milik user yang sedang login.
      parameters:
      - description: ID passkey
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Hapus passkey
      tags:
      - Passkeys
  /api/v1/passkeys/register:
    post:
      consumes:
      - application/json
      description: Memverifikasi respons attestation dan menyimpan passkey. Nama perangkat
        opsional via query `name`.
      parameters:
      - description: Nama perangkat
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Selesaikan registrasi passkey
      tags:
      - Passkeys
  /api/v1/passkeys/register/options:
    post:
      description: Mengembalikan opsi WebAuthn (publicKey) untuk membuat passkey baru;
        menyimpan challenge di cookie `ebitda_passkey`.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Mulai registrasi passkey
      tags:
      - Passkeys
...
  /api/v1/profile:
    patch:
      consumes:
      - application/json
      description: Mengubah nama dan email user yang sedang login.
      parameters:
      - description: Data profil
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.updateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Perbarui profil
      tags:
      - Profile
  /api/v1/profile/password:
    put:
      consumes:
      - application/json
      description: Mengganti kata sandi user yang sedang login (wajib kata sandi saat
        ini).
      parameters:
      - description: Data kata sandi
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/server.updatePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - CookieAuth: []
      summary: Ganti kata sandi
      tags:
      - Profile
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    in: cookie
    name: ebitda_session
    type: apiKey
swagger: "2.0"